Fix inaccurate color sample when at the extreme left of the canvas
When selecting a pure white color (#ffffff) the cursor can be
moved to the far left of the canvas. However the closest to white it
could ever get would be #fffefe.

It seems that the issue had to do with the linear gradient used to
display the current hue.

A linear gradient can be created with: ctx.createLinearGradient(x, y, ...)
where x == 0 is the far left of the color picker.  The actual value pulled
from the canvas at x == 0 with ctx.getImageData() is never actually
pure white though.

If the gradient is started one pixel offset to the right then a pure
white color can be selected: ctx.createLinearGradient(1, 0, ...)